more portal events, event values, tests
