--- src/test/skript/junit/ExprEvtPortal.sk
@@ -1,14 +1,0 @@
-options:
-	test: "org.skriptlang.skript.test.tests.syntaxes.expressions.ExprEvtPortalTest"
-
-test "ExprPortalTest" when running JUnit:
-	ensure {@test} completes "portal type - nether", "portal type - end" and "portal type - end gateway"
-
-on portal:
-	junit test is {@test}
-	if event-portaltype is nether portal:
-		complete objective "portal type - nether" for junit test {@test}
-	if event-portaltype is end portal:
-		complete objective "portal type - end" for junit test {@test}
-	if event-portaltype is end gateway:
-		complete objective "portal type - end gateway" for junit test {@test}
